misc: place both analytics version under the same url
