+++ src/pages/Home.tsx
@@ -1,16 +1,23 @@
 import { useEffect } from 'react'
-import { useNavigate } from 'react-router-dom'
+import { generatePath, useNavigate } from 'react-router-dom'
 
 import { getItemFromLS, removeItemFromLS } from '~/core/apolloClient'
-import { ANALYTIC_ROUTE, CUSTOMERS_LIST_ROUTE } from '~/core/router'
+import { NewAnalyticsTabsOptionsEnum } from '~/core/constants/tabsOptions'
+import { ANALYTIC_ROUTE, ANALYTIC_TABS_ROUTE, CUSTOMERS_LIST_ROUTE } from '~/core/router'
+import { PremiumIntegrationTypeEnum } from '~/generated/graphql'
 import { LAST_PRIVATE_VISITED_ROUTE_WHILE_NOT_CONNECTED_LS_KEY } from '~/hooks/core/useLocationHistory'
 import { useCurrentUser } from '~/hooks/useCurrentUser'
+import { useOrganizationInfos } from '~/hooks/useOrganizationInfos'
 import { usePermissions } from '~/hooks/usePermissions'
 
 const Home = () => {
   const navigate = useNavigate()
   const { loading: isUserLoading, currentMembership } = useCurrentUser()
   const { hasPermissions } = usePermissions()
+  const { hasOrganizationPremiumAddon } = useOrganizationInfos()
+  const hasAccessToRevenueAnalyticsFeature = hasOrganizationPremiumAddon(
+    PremiumIntegrationTypeEnum.RevenueAnalytics,
+  )
 
   useEffect(() => {
     // Make sure user permissions are loaded before performing redirection
@@ -26,8 +33,15 @@ const Home = () => {
         navigate(lastPrivateVisitedRouteWhileNotConnected, { replace: true })
         // This is a temp value for redirection, should be removed after redirection have been performed
         removeItemFromLS(LAST_PRIVATE_VISITED_ROUTE_WHILE_NOT_CONNECTED_LS_KEY)
-      } else if (hasPermissions(['analyticsView'])) {
+      } else if (hasPermissions(['analyticsView']) && !hasAccessToRevenueAnalyticsFeature) {
         navigate(ANALYTIC_ROUTE, { replace: true })
+      } else if (hasPermissions(['dataApiView']) && hasAccessToRevenueAnalyticsFeature) {
+        navigate(
+          generatePath(ANALYTIC_TABS_ROUTE, {
+            tab: NewAnalyticsTabsOptionsEnum.revenueStreams,
+          }),
+          { replace: true },
+        )
       } else {
         navigate(CUSTOMERS_LIST_ROUTE, { replace: true })
       }
