+++ src/pages/analytics/NewAnalytics.tsx
@@ -2,13 +2,9 @@ import { generatePath } from 'react-router-dom'
 
 import { NavigationTab, Typography } from '~/components/designSystem'
 import { NewAnalyticsTabsOptionsEnum } from '~/core/constants/tabsOptions'
-import { NEW_ANALYTIC_ROUTE, NEW_ANALYTIC_TABS_ROUTE } from '~/core/router'
+import { ANALYTIC_TABS_ROUTE } from '~/core/router'
 import { useInternationalization } from '~/hooks/core/useInternationalization'
-import Invoices from '~/pages/analytics/Invoices'
-import Mrr from '~/pages/analytics/Mrr'
-import PrepaidCredits from '~/pages/analytics/PrepaidCredits'
 import RevenueStreams from '~/pages/analytics/RevenueStreams'
-import Usage from '~/pages/analytics/Usage'
 import { PageHeader } from '~/styles'
 
 const NewAnalytics = () => {
@@ -27,44 +23,15 @@ const NewAnalytics = () => {
         tabs={[
           {
             title: translate('text_1739203651003n5f5qzxnhin'),
-            link: generatePath(NEW_ANALYTIC_TABS_ROUTE, {
+            link: generatePath(ANALYTIC_TABS_ROUTE, {
               tab: NewAnalyticsTabsOptionsEnum.revenueStreams,
             }),
             match: [
-              generatePath(NEW_ANALYTIC_ROUTE),
-              generatePath(NEW_ANALYTIC_TABS_ROUTE, {
+              generatePath(ANALYTIC_TABS_ROUTE, {
                 tab: NewAnalyticsTabsOptionsEnum.revenueStreams,
               }),
             ],
             component: <RevenueStreams />,
-          },
-          {
-            title: translate('text_6553885df387fd0097fd738c'),
-            link: generatePath(NEW_ANALYTIC_TABS_ROUTE, {
-              tab: NewAnalyticsTabsOptionsEnum.mrr,
-            }),
-            component: <Mrr />,
-          },
-          {
-            title: translate('text_17391997655892qv3sstq46n'),
-            link: generatePath(NEW_ANALYTIC_TABS_ROUTE, {
-              tab: NewAnalyticsTabsOptionsEnum.usage,
-            }),
-            component: <Usage />,
-          },
-          {
-            title: translate('text_637ccf8133d2c9a7d11ce6e1'),
-            link: generatePath(NEW_ANALYTIC_TABS_ROUTE, {
-              tab: NewAnalyticsTabsOptionsEnum.prepaidCredits,
-            }),
-            component: <PrepaidCredits />,
-          },
-          {
-            title: translate('text_63ac86d797f728a87b2f9f85'),
-            link: generatePath(NEW_ANALYTIC_TABS_ROUTE, {
-              tab: NewAnalyticsTabsOptionsEnum.invoices,
-            }),
-            component: <Invoices />,
           },
         ]}
       />
