+++ tailwind.config.ts
@@ -164,6 +164,7 @@ const config = {
         75: '18.75rem',
         76: '19rem',
         78: '19.5rem',
+        89: '22.25rem',
         90: '22.5rem',
         98: '24.5rem',
         100: '25rem',
