+++ translations/base.json
@@ -2809,7 +2809,6 @@
   "text_1738595318403vcyh77pwiew": "Partner billing",
   "text_1738605383523lme9aweoipp": "This invoice has been generated on behalf of this partner.",
   "text_1738748043939zqoqzz350yj": "Refresh",
-  "text_17391997655892qv3sstq46n": "Usage revenue",
   "text_1739804207625v7faokht1cq": "No lifetime usage has been reported",
   "text_1739804207625nkct7jhmmwh": "There was an error while fetching lifetime usage. Please try again.",
   "text_1737471851634wpeojigr27w": "Record a payment",
